src/backend/requirements.txt
absl-py==1.2.0
aiohttp==3.8.3
aiosignal==1.3.1
aniso8601==9.0.1
anyio==3.6.2
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.1.0
astunparse==1.6.3
async-timeout==4.0.2
asyncio==3.4.3
attrs==22.2.0
backcall==0.2.0
beautifulsoup4==4.11.1
bidict==0.22.1
bleach==5.0.1
cachetools==5.2.0
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
contourpy==1.0.5
cycler==0.11.0
debugpy==1.6.3
decorator==5.1.1
defusedxml==0.7.1
dnspython==2.3.0
entrypoints==0.4
eventlet==0.33.3
executing==1.2.0
fastjsonschema==2.16.2
filetype==1.0.7
Flask==2.2.2
Flask-Cors==3.0.10
Flask-RESTful==0.3.9
Flask-SocketIO==5.3.2
flatbuffers==22.10.26
fonttools==4.37.4
frozenlist==1.3.3
gast==0.4.0
google-auth==2.14.1
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
greenlet==2.0.1
grpcio==1.50.0
h5py==3.7.0
idna==3.4
imageio==2.25.0
importlib-metadata==6.0.0
imutils==0.5.4
ipykernel==6.17.1
ipython==8.6.0
ipython-genutils==0.2.0
ipywidgets==8.0.2
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.1
jupyter==1.0.0
jupyter-console==6.4.4
jupyter-server==1.23.3
jupyter_client==7.4.7
jupyter_core==5.0.0
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.3
keras==2.11.0
kiwisolver==1.4.4
libclang==14.0.6
Markdown==3.4.1
MarkupSafe==2.1.2
matplotlib==3.6.1
matplotlib-inline==0.1.6
mediapipe==0.8.9
mistune==2.0.4
multidict==6.0.4
nbclassic==0.4.8
nbclient==0.7.0
nbconvert==7.2.5
nbformat==5.7.0
nest-asyncio==1.5.6
notebook==6.5.2
notebook_shim==0.2.2
numpy==1.24.1
oauthlib==3.2.2
onnxruntime==1.9.0
opencv-contrib-python==4.6.0.66
opencv-python==4.7.0.68
opt-einsum==3.3.0
packaging==21.3
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==2.5.4
prometheus-client==0.15.0
prompt-toolkit==3.0.33
protobuf==3.19.6
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
Pygments==2.13.0
pyparsing==3.0.9
pyrsistent==0.19.2
python-dateutil==2.8.2
python-dotenv==0.21.1
python-engineio==4.3.4
python-socketio==5.7.2
pytz==2022.7.1
pyzmq==24.0.1
qtconsole==5.4.0
QtPy==2.3.0
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.0.1
scipy==1.9.3
Send2Trash==1.8.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.3.2.post1
stack-data==0.6.1
tensorboard==2.11.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-estimator==2.11.0
tensorflow-io-gcs-filesystem==0.28.0
termcolor==2.1.1
terminado==0.17.0
threadpoolctl==3.1.0
tinycss2==1.2.1
torch==1.13.0
torchvision==0.14.0
tornado==6.2
traitlets==5.5.0
typing_extensions==4.4.0
urllib3==1.26.14
wcwidth==0.2.5
webcolors==1.11.1
webencodings==0.5.1
websocket-client==1.4.2
Werkzeug==2.2.2
widgetsnbextension==4.0.3
wrapt==1.14.1
yarl==1.8.2
zipp==3.11.0
